backend/requirements.txt
fastapi
pydantic
sqlalchemy
asyncpg
uvicorn
email-validator
redis
minio
argon2-cffi
python-dotenv
urllib3
python-multipart
aiosmtplib
websockets